feat: split address component
